Refactor pyqwk into package with CLI and GUI
