+++ pyqwk/__init__.py
@@ -1,0 +1,6 @@
+"""Python QWK parsing library."""
+
+from pyqwk.core import *
+from pyqwk.core import __version__
+
+__all__ = [name for name in globals() if not name.startswith("_")]
