+++ tests/test_binary_detection.py
@@ -5,7 +5,7 @@ import pytest
 ROOT = Path(__file__).resolve().parents[1]
 sys.path.insert(0, str(ROOT))
 
-from qwk import _is_binary_line
+from pyqwk.core import _is_binary_line
 
 class TestBinaryDetection:
     """Unit tests for the _is_binary_line function."""
