+++ tests/test_input_validation.py
@@ -7,7 +7,7 @@ from pathlib import Path
 ROOT = Path(__file__).resolve().parents[1]
 sys.path.insert(0, str(ROOT))
 
-from qwk import (
+from pyqwk.core import (
     load_data,
     process_file,
     ProcessingSettings,
